remove solver registeration
