+++ pyomo/contrib/mindtpy/util.py
@@ -22,6 +22,7 @@ from pyomo.contrib.gdpopt.util import get_main_elapsed_time, time_code
 from pyomo.core.expr.calculus.derivatives import differentiate
 from pyomo.common.dependencies import attempt_import
 from pyomo.contrib.fbbt.fbbt import fbbt
+from pyomo.solvers.plugins.solvers.gurobi_persistent import GurobiPersistent
 
 pyomo_nlp = attempt_import('pyomo.contrib.pynumero.interfaces.pyomo_nlp')[0]
 numpy = attempt_import('numpy')[0]
@@ -606,3 +607,12 @@ def setup_solve_data(model, config):
                 direction=Suffix.IMPORT)
 
     return solve_data
+
+
+class GurobiPersistent4MindtPy(GurobiPersistent):
+
+    def _intermediate_callback(self):
+        def f(gurobi_model, where):
+            self._callback_func(self._pyomo_model, self,
+                                where, self.solve_data, self.config)
+        return f
